+++ AppLayer/Controllers/CategoryController.cs
@@ -1,41 +1,5 @@
-﻿//using BLL.DTOs;
-//using BLL.Services;
-//using DAL.EF;
-//using DAL.EF.Tables;
-//using Microsoft.AspNetCore.Mvc;
-
-//namespace AppLayer.Controllers
-//{
-//    public class CategoryController : Controller
-//    {
-//        private readonly CategoryService _categoryService;
-
-//        public CategoryController(CategoryService categoryService)
-//        {
-//            _categoryService = categoryService;
-//        }
-
-
-
-//    }
-//}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-using BLL.DTOs;
+﻿using BLL.DTOs;
 using BLL.Services;
-using DAL.EF.Tables;
 using Microsoft.AspNetCore.Mvc;
 
 namespace AppLayer.Controllers
@@ -49,24 +13,36 @@ namespace AppLayer.Controllers
             _categoryService = categoryService;
         }
 
-        // Show all categories
+        // Everyone can see categories
         public IActionResult Index()
         {
             var data = _categoryService.Get();
             return View(data);
         }
 
-        // Load create page
+        // ================= CREATE =================
+
         [HttpGet]
         public IActionResult Create()
         {
+            // ONLY ADMIN
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             return View();
         }
 
-        // Insert category
         [HttpPost]
         public IActionResult Create(CategoryDTO category)
         {
+            // ONLY ADMIN
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             if (ModelState.IsValid)
             {
                 bool result = _categoryService.Create(category);
@@ -80,34 +56,31 @@ namespace AppLayer.Controllers
             return View(category);
         }
 
+        // ================= EDIT =================
 
-
-
-
-
-
-
-
-
-
-
-
-
-        //....................................
-
-        // Load edit page
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            // ONLY ADMIN
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             var data = _categoryService.Get(id);
 
             return View(data);
         }
 
-        // Update category
         [HttpPost]
         public IActionResult Edit(CategoryDTO category)
         {
+            // ONLY ADMIN
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             if (ModelState.IsValid)
             {
                 var result = _categoryService.Update(category);
@@ -121,9 +94,16 @@ namespace AppLayer.Controllers
             return View(category);
         }
 
-        // Delete category
+        // ================= DELETE =================
+
         public IActionResult Delete(int id)
         {
+            // ONLY ADMIN
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
+
             _categoryService.Delete(id);
 
             return RedirectToAction("Index");
